FilterCommentBackend/requirements.txt
flask
fastapi
uvicorn
transformers
torch
Pillow
langchain-google-genai
python-dotenv